Update event gens to new schema, optimizations
