+++ lib/event_generators/bounty_events.rb
@@ -12,7 +12,7 @@ class BountyEvents
     Vulnerability.where(project: @project).each do |vuln|
       if has_bounty? vuln
         if valid_date? vuln, vuln.notes['bounty']['date']
-          e = Event.create!(
+          Event.create!(
             title: "Bounty awarded for #{vuln.cve}",
             event_type: 'bounty',
             date: vuln.notes['bounty']['date'],
@@ -20,10 +20,9 @@ class BountyEvents
             icon: 'monetization_on',
             color: '#25C322',
             notes: {},
+            vulnerability: vuln,
             start_hidden: false
           )
-          vuln.events << e
-          vuln.save
         end
       end
     end
