+++ lib/event_generators/fix_events.rb
@@ -8,12 +8,15 @@ class FixEvents
     @desc = Writing.event_article('fix')
   end
 
+  def fixes
+    Fix.joins(:vulnerability).where('vulnerabilities.project_id' => @project.id)
+  end
+
   def generate
+    logger = TimingLogger.new(STDOUT, 'data:events:fixes')
     events = []
-    vuln_events = []
-    Fix.joins(:vulnerability)
-       .where('vulnerabilities.project_id' => @project.id)
-       .each do |fix|
+    logger.info "Building model objects"
+    fixes.each do |fix|
       e = Event.new(
         title: "Fix for #{fix.vulnerability.cve}: #{fix.commit.title}",
         date: fix.commit.date,
@@ -22,16 +25,13 @@ class FixEvents
         icon: 'lock',
         color: '#9FE69F',
         notes: {},
+        vulnerability_id: fix.vulnerability_id,
         start_hidden: false
       )
       events << e
-      vuln_events << VulnerabilityEvent.new(
-        vulnerability_id: fix.vulnerability_id,
-        event: e
-      )
     end
+    logger.info "Importing events"
     Event.import events
-    VulnerabilityEvent.import vuln_events
   end
 
 end
