--- lib/event_generators/release_events.rb
@@ -1,25 +1,0 @@
-require_relative '../../app/helpers/application_helper'
-
-class ReleaseEvents
-  include ApplicationHelper
-
-  def initialize(project)
-    @project = project
-    @desc = Writing.event_article('release')
-  end
-
-  def generate
-    Release.all.each do |release|
-      Event.create!(
-        title: ':title:',
-        event_type: 'release',
-        date: release.date_released,
-        description: Writing.event_article('release'),
-        icon: 'stars',
-        color: '#D57BCE',
-        notes: {},
-        start_hidden: false
-      )
-    end
-  end
-end
