+++ lib/event_generators/same_cwe_events.rb
@@ -1,0 +1,65 @@
+require_relative '../../app/helpers/application_helper'
+require 'timing_logger'
+
+class SameCweEvents
+  include ApplicationHelper
+
+  def initialize(project)
+    @project = project
+    @cwe_title_template = "Another :CWE_NAME: vulnerability announced"
+    @cwe_desc_template = <<~EOSTR
+      :CVE: was reported, which is the same type of vulnerability as this one. They are both *:CWE_NAME:*.
+
+      See other cases of [:tags: :CWE_NAME:](/tags/:CWE_SLUG:).
+    EOSTR
+  end
+
+  def generate
+    logger = TimingLogger.new(STDOUT, 'data:events:similar')
+    events = []
+    query = <<~EOSQL
+    SELECT vt1.vulnerability_id AS vuln1_id,
+           vt2.vulnerability_id AS vuln2_id,
+           v2.announced AS announced,
+           v2.cve AS cve,
+           t1.name AS cwe,
+           t1.shortname AS tag_shortname
+    FROM vulnerability_tags AS vt1
+      INNER JOIN vulnerability_tags AS vt2 ON (vt1.tag_id = vt2.tag_id AND vt1.id <> vt2.id)
+      INNER JOIN vulnerabilities AS v2 ON (vt2.vulnerability_id = v2.id)
+      INNER JOIN tags AS t1 ON (t1.id = vt1.tag_id)
+      WHERE t1.name LIKE 'CWE%'
+      AND v2.project_id = '#{@project.id}'
+    EOSQL
+    logger.info 'Execute SQL query'
+    results = ActiveRecord::Base.connection.execute(query)
+    logger.info 'Build records for import'
+    results.each do |vuln|
+      e = {
+        event_type: 'same cwe',
+        title: Writing.interpolate_details(@cwe_title_template, {
+          ':CWE_NAME:' => friendly_cwe(vuln['cwe'])
+        }),
+        date:  vuln['announced'],
+        description: Writing.interpolate_details(@cwe_desc_template, {
+          ':CVE:' => vuln['cve'],
+          ':CWE_NAME:' => vuln['cwe'],
+          ':CWE_SLUG:' => vuln['tag_shortname'],
+        }),
+        icon: 'bug_report',
+        color: '#BB55FF',
+        start_hidden: false
+      }
+      events << e
+    end
+    logger.info 'Import events'
+    columns = [:event_type, :title, :date, :description,
+               :icon, :color, :start_hidden]
+    Event.import columns, events
+  end
+
+  def normalized_cwe_string(cwe)
+    cwe.to_s.gsub(/[^0-9]/,'').to_i
+  end
+
+end
