+++ lib/event_generators/vcc_events.rb
@@ -8,12 +8,15 @@ class VccEvents
     @desc = Writing.event_article('vcc')
   end
 
+  def vccs
+    Vcc.joins(:vulnerability).where('vulnerabilities.project_id' => @project.id)
+  end
+
   def generate
+    logger = TimingLogger.new(STDOUT, 'data:events:vccs')
     events = []
-    vuln_events = []
-    Vcc.joins(:vulnerability)
-       .where('vulnerabilities.project_id' => @project.id)
-       .each do |vcc|
+    logger.info "Building model objects"
+    vccs.each do |vcc|
       e = Event.new(
         title: "Vulnerability-contributing commit for #{vcc.vulnerability.cve}: #{vcc.commit.title}",
         date: vcc.commit.date,
@@ -25,12 +28,8 @@ class VccEvents
         start_hidden: false
       )
       events << e
-      vuln_events << VulnerabilityEvent.new(
-        vulnerability_id: vcc.vulnerability_id,
-        event: e
-      )
     end
+    logger.info "Importing events"
     Event.import events
-    VulnerabilityEvent.import vuln_events
   end
 end
