+++ lib/event_generators/vulnerability_events.rb
@@ -10,8 +10,9 @@ class VulnerabilityEvents
 
   def generate
     Vulnerability.where(project: @project).each do |v|
-        e = Event.create!(
+        Event.create!(
           date: v.announced,
+          vulnerability: v,
           event_type: 'announced',
           title: "Vulnerability Announced for #{v.cve}",
           description: @desc_template,
@@ -20,8 +21,6 @@ class VulnerabilityEvents
           notes: {},
           start_hidden: false
         )
-        v.events << e
-        v.save
     end
   end
