+++ lib/event_generators/weekly_report_events.rb
@@ -14,6 +14,7 @@ class WeeklyReportEvents
     @revert_desc = Writing.event_article('weekly-reverts')
     @owner_desc = Writing.event_article('weekly-ownership')
     @refactors_desc = Writing.event_article('weekly-refactors')
+
     @cooks_tag = Tag.find_by(shortname: 'cooks') || Tag.create!(
       name: 'Lesson: Too Many Cooks',
       shortname: 'cooks',
@@ -44,6 +45,7 @@ class WeeklyReportEvents
         num_devs = 0
         new_events = []
         json.each do |_week, report|
+          report['nice_date'] = nice_date(report['date'])
           new_events << overall_report_event(vuln_id, report)
           new_events << new_developers_event(vuln_id, report)
           new_events << ownership_change(vuln_id, report)
@@ -52,15 +54,14 @@ class WeeklyReportEvents
           num_devs += report['new_developers'].size
         end
         tags << too_many_cooks(vuln_id, num_devs) if num_devs >= 10
-        # vuln_events += new_events.compact.map do |e|
-        #   VulnerabilityEvent.new(event: e, vulnerability_id: vuln_id)
-        # end
         events += new_events.compact
       end
     end
     events.compact!
     logger.info "Inserting events"
-    Event.import events, validate: false
+    columns = [ :date, :event_type, :title, :description, :icon, :color,
+                :vulnerability_id, :start_hidden, ]
+    Event.import columns, events, validate: false
     logger.info "Inserting tag data"
     VulnerabilityTag.import tags, validate: false
     logger.info "Done with weeklies"
@@ -71,59 +72,36 @@ class WeeklyReportEvents
     churn = report['insertions'].to_i + report['deletions'].to_i
     num_devs = report['developers'].size.to_s
     drive_bys = report['drive_bys'] || []
-    file_list = report['files'].map { |f| " * `#{f}`" }.join("\n")
-    desc = @report_desc.gsub /:[\w_]+:/, {
+    desc = @report_desc.gsub /:[\w]+:/, {
       ':num_devs:' => num_devs,
-      ':nice_date:' => nice_date(report),
+      ':nice_date:' => report['nice_date'],
       ':num_commits:' => report['commits'].to_s,
       ':num_insertions:' => report['insertions'].to_s,
       ':num_deletions:' => report['deletions'].to_s,
       ':num_files:' => report['files'].size.to_s,
       ':num_drive_bys:' => drive_bys.size.to_s,
-      ':file_list:' => file_list,
     }
-    # For performance reasons, we're not providing details on each file
-    # files = Array.new
-    # report['files'].each do | file, index |
-    #   path = Filepath.find_by(filepath: file)
-    #   files.push({ file: file, slug: (path.nil? ? "" : path.slug) })
-    # end
 
-    # For performance reasons, we're not providing details on who each person was during the week.
-    # devs = Array.new
-    # report['developers'].each do | email |
-    #   dev = Developer.find_by(email: email)
-    #   devs.push(dev.nil? ? nil : dev.nickname)
-    # end
-
-    Event.new(
-      date: report['date'],
-      event_type: 'changes',
-      title: <<~EOS,
-        #{report['commits']} commits by #{num_devs} developers, #{churn} lines changed
-      EOS
-      description: desc,
-      icon: 'assignment',
-      color: '#bbbbbb',
-      vulnerability_id: vuln_id,
-      start_hidden: true
-    )
+    return { # Constructing 100k event objects is slow - use hashes instead
+        date: report['date'],
+        event_type: 'changes',
+        title: <<~EOS,
+          #{report['commits']} commits by #{num_devs} developers, #{churn} lines changed
+        EOS
+        description: desc,
+        icon: 'assignment',
+        color: compute_color(churn),
+        vulnerability_id: vuln_id,
+        start_hidden: true
+      }
   end
 
   def new_developers_event(vuln_id, report)
     return nil unless report['new_developers'].any?
     num_new_devs = report['new_developers'].size
-    desc = @newdev_desc.gsub(':nice_date:', nice_date(report))
+    desc = @newdev_desc.gsub(':nice_date:', report['nice_date'])
                        .gsub(':new_devs:', num_new_devs.to_s)
-
-    # For performance reasons, we're not loading who the developer actually is here.
-    devs = Array.new
-    # report['new_developers'].each do | email |
-    #   dev = Developer.find_by(email: email)
-    #   devs.push(dev.nil? ? nil : dev.nickname)
-    # end
-
-    Event.new(
+    return {
       date: report['date'],
       event_type: 'new developer',
       title: "#{num_new_devs} new developer(s)",
@@ -132,13 +110,13 @@ class WeeklyReportEvents
       color: '#B7AADE',
       vulnerability_id: vuln_id,
       start_hidden: true
-    )
+    }
   end
 
   def ownership_change(vuln_id, report)
     return nil unless report['ownership_change']
-    desc = @owner_desc.gsub(':nice_date:', nice_date(report))
-    Event.new(
+    desc = @owner_desc.gsub(':nice_date:', report['nice_date'])
+    return {
       date: report['date'],
       title: "Code ownership changed",
       event_type: 'owner change',
@@ -147,15 +125,15 @@ class WeeklyReportEvents
       color: '#BB5093',
       vulnerability_id: vuln_id,
       start_hidden: false
-    )
+    }
   end
 
   def revert(vuln_id, report)
     return nil unless report['reverts'].size > 0
     desc = @revert_desc.gsub(':num_reverts:', report['reverts'].size.to_s)
-                       .gsub(':nice_date:', nice_date(report))
+                       .gsub(':nice_date:', report['nice_date'])
 
-    Event.new(
+    return {
       date: report['date'],
       title: "#{report['reverts'].size.to_s} revert commit(s)",
       event_type: 'reverts',
@@ -164,14 +142,14 @@ class WeeklyReportEvents
       color: '#FF953D',
       vulnerability_id: vuln_id,
       start_hidden: false
-    )
+    }
   end
 
   def refactor(vuln_id, report)
     return nil unless report['refactors'].size > 0
     desc = @refactors_desc.gsub(':num_refactors:', report['refactors'].size.to_s)
-                          .gsub(':nice_date:', nice_date(report))
-    Event.new(
+                          .gsub(':nice_date:', report['nice_date'])
+    return {
       date: report['date'],
       title: "#{report['refactors'].size.to_s} refactor commit(s)",
       event_type: 'refactors',
@@ -180,7 +158,7 @@ class WeeklyReportEvents
       color: '#F4DC0C',
       vulnerability_id: vuln_id,
       start_hidden: false
-    )
+    }
   end
 
   def too_many_cooks(vuln_id, num_devs)
@@ -191,14 +169,21 @@ class WeeklyReportEvents
     )
   end
 
-  def nice_date(report)
-    DateTime.parse(report['date']).strftime('%A, %B %d, %Y')
+  def nice_date(date_str)
+    DateTime.parse(date_str).strftime('%A, %B %d, %Y')
   end
 
   # Map every CVE to its ID - save us a DB call over and over
   def build_vuln_cache()
     r = Vulnerability.select(:id,:cve).where(project: @project).pluck(:id, :cve)
     return r.inject({}) { |map, (id, cve)| map[cve] = id; map }
+  end
+
+  def compute_color(churn)
+    big_week = 1_000.0 # >1k SLOC changed? That's a big week.
+    churn = churn.clamp(0, big_week)
+    perc = churn / big_week + 0.05 # don't want pure white on zero churn
+    return "hsl(0,0,#{perc}%)"
   end
 
 end
