+++ lib/tasks/chromium.rake
@@ -38,7 +38,7 @@ namespace :data do
       'data:no_logger',
       'data:chromium:load',
       'data:chromium:events',
-      # 'data:chromium:tags',
+      'data:chromium:tags',
       'data:optimize',
       'data:clear_cache',
     ]
@@ -94,12 +94,12 @@ namespace :data do
     task :events => :environment do
       project = Project.find_by(subdomain: 'chromium')
       [
-        # SimilarEvents,
-        # VulnerabilityEvents,
-        # FixEvents,
-        # VccEvents,
-        # ReleaseEvents,
-        # BountyEvents,
+        VulnerabilityEvents,
+        SameCweEvents,
+        SameDirectoryEvents,
+        FixEvents,
+        VccEvents,
+        BountyEvents
       ].each do |generator|
         logger.info "Generating #{generator.to_s.tableize.humanize.downcase}"
         generator.new(project).generate
