+++ lib/tasks/struts.rake
@@ -77,18 +77,18 @@ namespace :data do
     task :events => :environment do
       project = Project.find_by(subdomain: 'struts')
       [
-        SimilarEvents,
         VulnerabilityEvents,
+        SameCweEvents,
+        SameDirectoryEvents,
         FixEvents,
         VccEvents,
-        # ReleaseEvents,
         BountyEvents,
       ].each do |generator|
         logger.info "Generating #{generator.to_s.tableize.humanize.downcase}"
         generator.new(project).generate
       end
       logger.info "Generating weekly report events and tags"
-      WeeklyReportEvents.new.generate(repo_dir)
+      WeeklyReportEvents.new(project).generate(repo_dir)
     end
 
     desc 'Generate and apply tags'
