+++ lib/writing.rb
@@ -37,4 +37,10 @@ class Writing
     return new_str.join
   end
 
+  # This is a more updated version of embed_details - caller customizes the embeds.
+  # Only requirement is that the keys are :like_this:
+  def self.interpolate_details(template, embeds)
+     template.gsub /:[\w]+:/, embeds
+  end
+
 end
